refactor(rpc): move the compatibility fix to a separate module
